Added support for Doubles
Using USE_DOUBLE_PRECISION will enable Double Precision
Float performance should be unaffected

All Tests Pass

+++ src/JitterTests/BoundingBoxTests.cs
@@ -14,7 +14,7 @@ public class BoundingBoxTests
         ShapeHelper.CalculateBoundingBox(shape, ori, pos, out JBBox shr);
         shape.CalculateBoundingBox(ori, pos, out JBBox sbb);
 
-        float fraction = shr.GetVolume() / sbb.GetVolume();
+        double fraction = shr.GetVolume() / sbb.GetVolume();
 
         Assert.That(fraction - 1e-7f, Is.LessThan(1.0f));
         Assert.That(fraction, Is.GreaterThan(0.2f));
